requirements.txt
amqp==2.1.1
astroid==1.4.8
asyncio==3.4.3
attrs==16.0.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.4.1
billiard==3.5.0.2
bokeh==0.12.1
celery==4.0.0
cffi==1.7.0
chatterbot==0.4.5
click==6.6
colorama==0.3.7
cryptography==1.4
cssselect==0.9.2
cycler==0.10.0
decorator==4.0.10
Django==1.8.2
django-admin-bootstrapped==2.5.7
django-admin-bootstrapped-plus==0.1.1.2
django-bootstrap3==7.0.1
django-pagedown==0.1.1
dominate==2.3.1
ecdsa==0.13
entrypoints==0.2.2
et-xmlfile==1.0.1
fake-useragent==0.1.2
feedparser==5.2.1
Flask==0.11.1
Flask-Bootstrap==3.3.7.0
Flask-Login==0.3.2
flask-misaka==0.4.1
Flask-Moment==0.5.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
future==0.15.2
fuzzywuzzy==0.11.1
gevent==1.1.2
greenlet==0.4.10
html2text==2016.5.29
idna==2.1
ipykernel==4.3.1
ipython==4.2.0
ipython-genutils==0.1.0
isort==4.2.5
itchat==1.1.19
itsdangerous==0.24
jdcal==1.2
jedi==0.9.0
jieba==0.38
Jinja2==2.8
jsondatabase==0.1.1
jsonschema==2.5.1
jupyter-client==4.3.0
jupyter-core==4.1.0
kombu==4.0.0
lazy-object-proxy==1.2.2
lxml==3.6.0
markdown2==2.3.1
MarkupSafe==0.23
MathJaxPlugin==0.1.4
matplotlib==1.5.3
mccabe==0.5.2
mechanize==0.2.5
mistune==0.7.2
mongoengine==0.10.6
nbconvert==4.2.0
nbformat==4.0.1
nltk==3.2.1
notebook==4.2.1
numpy==1.11.2+mkl
oauthlib==1.1.2
openpyxl==2.3.5
panda==0.3.1
pandas==0.18.1
parsel==1.0.2
pdfminer3k==1.3.0
pep8==1.7.0
pickleshare==0.7.2
Pillow==3.3.1
ply==3.8
prettytable==0.7.2
py==1.4.31
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
PyDispatcher==2.0.5
pygal==2.2.3
pygame==1.9.2a0
pygeoip==0.3.2
Pygments==2.1.3
pylint==1.6.4
pymongo==3.3.0
PyMySQL==0.7.4
pyOpenSSL==16.0.0
pyparsing==2.1.5
pyPdf==1.13
pypiwin32==219
pyreadline==2.1
pytesser3==1.0.1
pytesseract==0.1.6
pytest==2.9.2
python-dateutil==2.5.3
python-nmap==0.6.0
python-twitter==3.1
pytz==2016.4
PyYAML==3.11
pyzmq==15.2.0
queuelib==1.4.2
redis==2.10.5
requests==2.12.1
requests-oauthlib==0.6.2
rsa==3.4.2
scikit-learn==0.18.1
scipy==0.18.1
Scrapy==1.1.0
selenium==2.53.6
service-identity==16.0.0
sgmllib3k==1.0.0
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
SQLAlchemy==1.0.13
textblob==0.11.1
tornado==4.3
traitlets==4.2.1
Twisted==16.2.0
twisted-win==0.5.5
urwid==1.3.1
vine==1.1.3
visitor==0.1.3
w3lib==1.14.2
websockets==3.2
Werkzeug==0.11.10
Whoosh==2.7.4
wrapt==1.10.8
WTForms==2.1
xlrd==1.0.0
zhihu-oauth==0.0.20
zhihu-py3==0.3.17
zope.interface==4.2.0
